Update project set up
