+++ .swiftlint.yml
@@ -14,7 +14,7 @@ opt_in_rules:
   - redundant_nil_coalesing
 #  - switch_case_on_newline # broken
 
-line_length: 110 
+line_length: 120 
   
 included:
   - Source
