+++ Source/Factories/ClusterFactory.swift
@@ -1,0 +1,18 @@
+//
+//  ClusterFactory.swift
+//  Matrioska
+//
+//  Created by Joan Romano on 13/01/17.
+//  Copyright © 2017 runtastic. All rights reserved.
+//
+
+import Foundation
+
+/// The standard library factories
+/// 1. `TabBarClusterFactory`: A tabBar cluster factory
+/// 2. `StackClusterFactory`: A stack cluster factory
+///
+/// See the documentation of the factories for more informations and configuration options.
+public enum ClusterFactory {
+    /* empty ATM, extended in other files... at some point there will be something here */
+}
